# Follow up on post-merge review of performance work on sharded searcher