This adresses the comments by @keegancsmith in https://github.com/sourcegraph/zoekt/pull/14 after the merge.

I'm quoting the commit message of the first commit here:

> 
> This commit does a couple of things to make the added behavior correct:
> 
> 1. It reuses the simplified query produced by `selectRepoSet` which
>    in turn reduces the work the shards have to do when evaluating the
>    query
> 2. It only evaluates the `query.RepoSet` if it's part of a top-level
>    `query.And`. The previous code was too naive when considering that a
>    `query.RepoSet` can appear anywhere in a `query.Q`. See this comment
>    for more details:
>    #14 (comment)
> 3. It changes the const value that's returned in the simplified version
>    so that `typeRepoSearcher`, which evaluates `query.TypeRepo` queries to
>    a `query.RepoSet` works (the tests in `shards/eval_test.go` break
>    without this.

I think the behavior is now correct.

I don't think it's worth it, but if the usecase of multiple RepoSet queries ever arises we can easily remove the `break` and use a set to not append duplicates to the `filtered` slice.