@mrnugget 2019-07-20
Hmm, I can take a look again, but when I tried that (returning the
simplified version unconditionally) the tests for TypeRepo broke because
it’s using a RepoSet.

On Sat 20. Jul 2019 at 15:17, Keegan Carruthers-Smith <
notifications@github.com> wrote:

> *@keegancsmith* approved this pull request.
> ------------------------------
>
> In shards/shards.go
> <https://github.com/sourcegraph/zoekt/pull/15#discussion_r305579938>:
>
> >
> -		return &query.Const{Value: true}
> -	})
> -	query.Simplify(eval)
> -
> -	if len(filtered) != 0 {
> -		return filtered
> +		// Stop after first RepoSet, otherwise we might append duplicate
> +		// shards to `filtered`
> +		if len(filtered) != 0 {
> +			return filtered, and
>
> I think you can just always return here, don't have to check len of
> filtered.
>
> Also return query.Simplify(and), to make Simplify do nothing per shard.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/sourcegraph/zoekt/pull/15?email_source=notifications&email_token=AAJBLZN5OIQ3MVVMV4S3SCDQAMF63A5CNFSM4IFGSJ2KYY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOB7B6POY#pullrequestreview-264497083>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAJBLZIKSC5ZGXBRVOMZ4ZLQAMF63ANCNFSM4IFGSJ2A>
> .
>