@keegancsmith 2019-07-22
Tried out add Simplify, and the tests pass. Also the benchmarks show an improvement:

```
name                                                 old time/op    new time/op    delta
ShardedSearch/substring_all_results-8                   1.54s ±12%     1.48s ± 8%     ~     (p=0.421 n=5+5)
ShardedSearch/substring_no_results-8                   5.43ms ± 2%    5.31ms ± 5%     ~     (p=0.548 n=5+5)
ShardedSearch/substring_some_results-8                 11.3ms ± 2%    10.6ms ± 3%   -5.99%  (p=0.008 n=5+5)
ShardedSearch/substring_all_results_and_repo_set-8      665ms ± 3%     659ms ± 3%     ~     (p=0.421 n=5+5)
ShardedSearch/substring_some_results_and_repo_set-8    6.21ms ± 4%    6.10ms ± 6%     ~     (p=0.690 n=5+5)
ShardedSearch/substring_no_results_and_repo_set-8      3.49ms ± 4%    3.25ms ± 3%   -6.96%  (p=0.008 n=5+5)

name                                                 old alloc/op   new alloc/op   delta
ShardedSearch/substring_all_results-8                  2.98GB ± 0%    2.98GB ± 0%     ~     (p=1.000 n=5+5)
ShardedSearch/substring_no_results-8                   8.35MB ± 0%    8.35MB ± 0%     ~     (p=0.286 n=4+5)
ShardedSearch/substring_some_results-8                 18.4MB ± 0%    18.4MB ± 0%     ~     (p=1.000 n=5+5)
ShardedSearch/substring_all_results_and_repo_set-8     1.50GB ± 0%    1.50GB ± 0%   -0.01%  (p=0.008 n=5+5)
ShardedSearch/substring_some_results_and_repo_set-8    9.32MB ± 0%    9.13MB ± 0%   -2.06%  (p=0.016 n=5+4)
ShardedSearch/substring_no_results_and_repo_set-8      4.59MB ± 0%    4.40MB ± 0%   -4.18%  (p=0.008 n=5+5)

name                                                 old allocs/op  new allocs/op  delta
ShardedSearch/substring_all_results-8                   27.2M ± 0%     27.2M ± 0%     ~     (p=0.686 n=4+4)
ShardedSearch/substring_no_results-8                     105k ± 0%      105k ± 0%     ~     (all equal)
ShardedSearch/substring_some_results-8                   240k ± 0%      240k ± 0%     ~     (p=1.000 n=4+5)
ShardedSearch/substring_all_results_and_repo_set-8      13.6M ± 0%     13.6M ± 0%   -0.04%  (p=0.008 n=5+5)
ShardedSearch/substring_some_results_and_repo_set-8      126k ± 0%      120k ± 0%   -4.75%  (p=0.000 n=5+4)
ShardedSearch/substring_no_results_and_repo_set-8       58.6k ± 0%     52.6k ± 0%  -10.24%  (p=0.008 n=5+5)
```