@mrnugget 2019-07-23
Ah, sweet! Thanks for that! I guess I had a combination of the previous version and this that didn't work.